Remove url pattern parsing until rules are implemented. Cleanup python files
